Move DataItemBase into a separate file
