+++ ios/StatusPanel/Data/DataItemBase.swift
@@ -1,0 +1,49 @@
+// Copyright (c) 2018-2023 Jason Morley, Tom Sutcliffe
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+import Foundation
+import UIKit
+
+protocol DataItemBase : AnyObject {
+
+    var icon: String? { get }
+    var prefix: String { get }
+    var flags: DataItemFlags { get }
+    var subText: String? { get }
+    var accentColor: UIColor? { get }
+
+    func getText(checkFit: (String) -> Bool) -> String
+}
+
+extension DataItemBase {
+
+    var iconAndPrefix: String {
+        var elements: [String] = []
+        if let icon = self.icon {
+            elements.append(icon)
+        }
+        let prefix = self.prefix
+        if !prefix.isEmpty {
+            elements.append(prefix)
+        }
+        return elements.joined(separator: " ")
+    }
+
+}
